feat: implemented hashing for password persistence and validation
